pyproject.toml
[project]
name = "coursework-sql-andrianov-m-s"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksim Andreev",email = "andrianov_maksim@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
mypy = "^1.15.0"
isort = "^6.0.1"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"



[tool.mypy]
disallow_untyped_defs = false
no_implicit_optional = false
warn_return_any = true
warn_unreachable = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.gitignore
    | dist
  )/
)
'''

[tool.isort]
line_length = 119
skip_glob = ["docs/*"]
skip = [".gitignore", ".dockerignore"]
